Activate per-dataset feature flags when the property is set
If a compression algorithm requires a feature flag, activate the
feature when the property is set, rather than waiting for the first
block to be born.  This is currently only used by zstd but can be
extended as needed.

Return -1 from zfs_prop_set_special() as we only want to handle
feature activation in zfs_prop_set_special(), the actual property
setting can still happen the normal way.

Sponsored-by: The FreeBSD Foundation
Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>
Signed-off-by: Allan Jude <allanjude@freebsd.org>